requirements.txt
boto3==1.24.12
Django==4.1.3
django-environ==0.11.2